refactor: simplified setup API
It is no longer necessary to set initialProperties, and the library
exports a static `application:continueUserActivity:restorationHandler:` handler.

The listener is also no longer called immediately with the initial shortcut.

Instead, the library now provides a `getInitialShortcut` async method to get the
initial shortcut. This should make it easier to integrate with `react-navigation`'s
Deep Linking.

We should also write a React Navigation Example at some point, with Deep Linking setup.

A migration guide needs to be written for this change. This release WILL BE A BREAKING CHANGE.